Move swipe gesture handling upwards
